[#3] refactor: modularize signup/login components
